EPMECO-12860. Make schema compatibility check with global param.
Update io.confluent libraries to version 7.1.1.

Deleted unnecessary changes.

+++ commons/src/main/java/com/epam/eco/schemacatalog/client/CachedExtendedSchemaRegistryClient.java
@@ -573,7 +573,7 @@ public class CachedExtendedSchemaRegistryClient extends EcoCachedSchemaRegistryC
             String subject,
             Integer version) throws IOException, RestClientException {
         if (version != null) {
-            return getSchemaMetadata(subject, version);
+            return getSchemaMetadata(subject, version.intValue());
         } else {
             return getLatestSchemaMetadata(subject);
         }
